@@ -1,7 +1,7 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 28
+    compileSdkVersion 29
 
     /*
     V7: Android 2.1.x	ECLAIR_MR1
@@ -23,9 +23,9 @@ android {
     defaultConfig {
         applicationId "de.euhm.jlt"
         minSdkVersion 14
-        targetSdkVersion 28
-        versionCode 13
-        versionName "0.15"
+        targetSdkVersion 29
+        versionCode 14
+        versionName "0.16"
     }
 
     signingConfigs {
@@ -50,9 +50,9 @@ android {
 }
 
 dependencies {
-    implementation 'com.android.support:support-v4:28.0.0'
-    implementation 'com.android.support:appcompat-v7:28.0.0'
-    implementation 'com.android.support:design:28.0.0'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'androidx.appcompat:appcompat:1.0.0'
+    implementation 'com.google.android.material:material:1.0.0'
 }
 
 System.setProperty('java.awt.headless', 'false')
