@@ -7,10 +7,11 @@
 package de.euhm.jlt;
 
 import android.content.Context;
-import android.support.v4.view.ViewPager;
 import android.util.AttributeSet;
 import android.view.MotionEvent;
 import android.view.View;
+
+import androidx.viewpager.widget.ViewPager;
 
 /**
  * Custom ViewPager to catch swipe event within statistics
