@@ -8,8 +8,6 @@
  */
 package de.euhm.jlt.dialogs;
 
-import java.util.Calendar;
-
 import android.annotation.SuppressLint;
 import android.annotation.TargetApi;
 import android.app.Activity;
@@ -18,11 +16,15 @@ import android.app.Dialog;
 import android.content.DialogInterface;
 import android.os.Build;
 import android.os.Bundle;
-import android.support.v4.app.DialogFragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.widget.DatePicker;
 import android.widget.TimePicker;
+
+import androidx.fragment.app.DialogFragment;
+
+import java.util.Calendar;
+
 import de.euhm.jlt.R;
 import de.euhm.jlt.dao.Times;
 import de.euhm.jlt.utils.Constants;
