@@ -7,19 +7,21 @@
  */
 package de.euhm.jlt.dialogs;
 
-import java.util.Calendar;
-import java.util.Locale;
-
-import android.content.Context;
 import android.app.AlertDialog;
 import android.app.Dialog;
+import android.content.Context;
 import android.content.DialogInterface;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.v4.app.DialogFragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.widget.NumberPicker;
+
+import androidx.annotation.NonNull;
+import androidx.fragment.app.DialogFragment;
+
+import java.util.Calendar;
+import java.util.Locale;
+
 import de.euhm.jlt.R;
 import de.euhm.jlt.utils.Constants;
 
