@@ -6,15 +6,16 @@
  */
 package de.euhm.jlt.dialogs;
 
-import java.util.Calendar;
-
 import android.app.Activity;
 import android.app.Dialog;
 import android.app.TimePickerDialog;
 import android.os.Bundle;
-import android.support.v4.app.DialogFragment;
 import android.text.format.DateFormat;
 import android.widget.TimePicker;
+
+import androidx.fragment.app.DialogFragment;
+
+import java.util.Calendar;
 
 /**
  * Implements a DialogFragment for picking the time
