@@ -6,9 +6,6 @@
  */
 package de.euhm.jlt.preferences;
 
-import java.lang.reflect.Constructor;
-import java.lang.reflect.Method;
-
 import android.annotation.SuppressLint;
 import android.app.Activity;
 import android.content.Context;
@@ -20,11 +17,15 @@ import android.os.Message;
 import android.preference.Preference;
 import android.preference.PreferenceManager;
 import android.preference.PreferenceScreen;
-import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ListView;
+
+import androidx.fragment.app.Fragment;
+
+import java.lang.reflect.Constructor;
+import java.lang.reflect.Method;
 
 /**
  * A PreferenceFragment for the support library. Based on the platform's code 
